Update patch set 4
Patch Set 4:

I know the approach to solve it, just not how to do it.

/sys/class/leds/led:flash_torch/brightness
has permission 660 (rw-rw----). This should be 666 (rw-rw-rw-). If I change the permission to 666, Torch works.

So the permission of the sys parameter needs to be changed but I don't know how to do this.

Patch-set: 4
Label: Code-Review=0