hammerhead: Clean up Torch Two squashed commits
hammerhead: Remove global-read/write for torch sysfs

This is a security issue - anything can turn on the LED right now

This is not the proper way to fix torch.

hammerhead: Make OmniTorch use camera interface

Even if you solve permissions issues with sysfs, Torch causes
the camera to be unable to control the flash if you use Torch
then switch to camera when using sysfs.

This has been reported by multiple users of both Omni and CM.

When using the camera interface, Torch works without permissions hacks and does not interfere with normal camera operation.

Change-Id: I4eb005cb7fce6908a4cfb5fbc7b3e575bd6ab308