hammerhead: Remove global-read/write for torch sysfs
This is a security issue - anything can turn on the LED right now

This is not the proper way to fix torch.  Even if changing this
breaks torch for now, we can't use this approach for security reasons.

So if torch is broken, it will stay broken until we figure out how to
make it part of the camera group.

Change-Id: I4eb005cb7fce6908a4cfb5fbc7b3e575bd6ab308