Update patch set 1
Patch Set 1:

Well, first I need to see if we really need most of this things, and if we need them I will look if there are separated commits for this. is not good to do it via merge since is better to have all commits separated so we can revert in case it is needed or for maintain authorship correctly. I will keep an eye on this when I get my sources running good, seems Google has some connection troubles.

Patch-set: 1