Update patch set 5
Patch Set 5:

It may be a no-op in master, but this is not master.

Although it looks like it might be a no-op here as build/core/product.mk seems to set this to true unless it's forced false.

Patch-set: 5
Reviewer: Gerrit User 1000008 <1000008@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=0